Package UI fxml and update javafx version
