+++ src/main/java/seedu/address/ui/group/GroupCard.java
@@ -1,4 +1,6 @@
-package seedu.address.ui;
+package seedu.address.ui.group;
+
+import static java.util.stream.Collectors.joining;
 
 import java.util.Comparator;
 
@@ -7,14 +9,18 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
+import seedu.address.model.group.Group;
+import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
+import seedu.address.ui.UiPart;
+
 
 /**
  * An UI component that displays information of a {@code Person}.
  */
-public class PersonCard extends UiPart<Region> {
+public class GroupCard extends UiPart<Region> {
 
-    private static final String FXML = "PersonListCard.fxml";
+    private static final String FXML = "GroupListCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -24,38 +30,32 @@ public class PersonCard extends UiPart<Region> {
      * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on AddressBook level 4</a>
      */
 
-    public final Person person;
+    public final Group group;
 
     @FXML
     private HBox cardPane;
     @FXML
-    private Label name;
+    private Label groupName;
     @FXML
     private Label id;
     @FXML
-    private Label phone;
-    @FXML
-    private Label address;
-    @FXML
-    private Label email;
-    @FXML
-    private Label groups;
+    private Label members;
     @FXML
     private FlowPane tags;
 
     /**
-     * Creates a {@code PersonCode} with the given {@code Person} and index to display.
+     * Creates a {@code GroupCode} with the given {@code Group} and index to display.
      */
-    public PersonCard(Person person, int displayedIndex) {
+    public GroupCard(Group group, int displayedIndex) {
         super(FXML);
-        this.person = person;
+        this.group = group;
         id.setText(displayedIndex + ". ");
-        name.setText(person.getName().fullName);
-        phone.setText(person.getPhone().value);
-        address.setText(person.getAddress().value);
-        email.setText(person.getEmail().value);
-        groups.setText(person.getGroupsName());
-        person.getTags().stream()
+        groupName.setText(group.getName());
+        members.setText(group.getMembers().stream()
+                .map(Person::getName)
+                .map(Name::toString)
+                .collect(joining(", ")));
+        group.getTags().stream()
                 .sorted(Comparator.comparing(tag -> tag.tagName))
                 .forEach(tag -> tags.getChildren().add(new Label(tag.tagName)));
     }
@@ -68,13 +68,13 @@ public class PersonCard extends UiPart<Region> {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof PersonCard)) {
+        if (!(other instanceof GroupCard)) {
             return false;
         }
 
         // state check
-        PersonCard card = (PersonCard) other;
+        GroupCard card = (GroupCard) other;
         return id.getText().equals(card.id.getText())
-                && person.equals(card.person);
+                && group.equals(card.group);
     }
 }
