+++ src/main/java/seedu/address/ui/person/PersonCard.java
@@ -1,6 +1,4 @@
-package seedu.address.ui;
-
-import static java.util.stream.Collectors.joining;
+package seedu.address.ui.person;
 
 import java.util.Comparator;
 
@@ -9,17 +7,15 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.group.Group;
-import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
-
+import seedu.address.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Person}.
  */
-public class GroupCard extends UiPart<Region> {
+public class PersonCard extends UiPart<Region> {
 
-    private static final String FXML = "GroupListCard.fxml";
+    private static final String FXML = "PersonListCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -29,32 +25,38 @@ public class GroupCard extends UiPart<Region> {
      * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on AddressBook level 4</a>
      */
 
-    public final Group group;
+    public final Person person;
 
     @FXML
     private HBox cardPane;
     @FXML
-    private Label groupName;
+    private Label name;
     @FXML
     private Label id;
     @FXML
-    private Label members;
+    private Label phone;
+    @FXML
+    private Label address;
+    @FXML
+    private Label email;
+    @FXML
+    private Label groups;
     @FXML
     private FlowPane tags;
 
     /**
-     * Creates a {@code GroupCode} with the given {@code Group} and index to display.
+     * Creates a {@code PersonCode} with the given {@code Person} and index to display.
      */
-    public GroupCard(Group group, int displayedIndex) {
+    public PersonCard(Person person, int displayedIndex) {
         super(FXML);
-        this.group = group;
+        this.person = person;
         id.setText(displayedIndex + ". ");
-        groupName.setText(group.getName());
-        members.setText(group.getMembers().stream()
-                .map(Person::getName)
-                .map(Name::toString)
-                .collect(joining(", ")));
-        group.getTags().stream()
+        name.setText(person.getName().fullName);
+        phone.setText(person.getPhone().value);
+        address.setText(person.getAddress().value);
+        email.setText(person.getEmail().value);
+        groups.setText(person.getGroupsName());
+        person.getTags().stream()
                 .sorted(Comparator.comparing(tag -> tag.tagName))
                 .forEach(tag -> tags.getChildren().add(new Label(tag.tagName)));
     }
@@ -67,13 +69,13 @@ public class GroupCard extends UiPart<Region> {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof GroupCard)) {
+        if (!(other instanceof PersonCard)) {
             return false;
         }
 
         // state check
-        GroupCard card = (GroupCard) other;
+        PersonCard card = (PersonCard) other;
         return id.getText().equals(card.id.getText())
-                && group.equals(card.group);
+                && person.equals(card.person);
     }
 }
