+++ src/main/resources/view/ViewPanel.fxml
@@ -5,7 +5,7 @@
 
 <?import javafx.geometry.Insets?>
 <?import java.net.URL?>
-<StackPane xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1">
+<StackPane xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
     <stylesheets>
         <URL value="@DarkTheme.css" />
     </stylesheets>
